26.06.2026

+++ Assets/Scriptable Object/Data Scripts/Entities/PlayerScriptData.cs
@@ -3,9 +3,9 @@
 [CreateAssetMenu(fileName = "PlayerStatBlock", menuName = "StatBlock/Player Stat Block", order = -1000)]
 public class PlayerScriptData : ScriptableObject
 {
-    public int HP;
-    public int Stamina;
-    public int WalkingSpeed;
-    public int RunSpeed;
-    public int SlowSpeed;
+    public float HP;
+    public float Stamina;
+    public float WalkingSpeed;
+    public float RunSpeed;
+    public float SlowSpeed;
 }
